Both server and client can both send multiple msgs
Before, the client sent and received one message in that order. So even
though the server could theoretically send and receive more than one
message in a row, the client wouldn't read more than one message in a
row (because the client would wait on a user to write a message to send
before reading the next message) and the client wouldn't send more than
one message in a row because it would immediately make a blocking call
to recv to read another message

Both the client and server have to be capable of handling more than one
message in a row to actually see the results of either program being
able to concurrently accept and receive messages.