Allow convo history to print to a file
This makes a quick-and-dirty TUI for chatting. The terminal where the
program is running is where user input is entered. You can provide
another file where the conversation will be written. This file can be a
tty so you can view the conversation in another terminal, which allows
for a quick-and-dirty TUI.

Changes to be committed:
      modified:   chat.py
      modified:   client-chat.py
      modified:   my_socket.py
      modified:   notes.md